requirements.txt
# Core
#torch>=2.0.0
#conda install pytorch==2.3.1 torchvision==0.18.1 torchaudio==2.3.1 pytorch-cuda=12.1 -c pytorch -c nvidia

numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Processing
duckdb>=0.9.0
pyarrow>=14.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Explainability
shap>=0.43.0

# Progress
tqdm>=4.65.0

# Config
pyyaml>=6.0

# Testing
pytest>=7.0.0
